Improve information display for Human Player
Deduplicate and sort hinted information about held cards.
Instead of displaying:

Card 1 is b but is not 122
Card 2 is 22 but is not 1b
Card 3 is unknown but is not 1b22
Card 4 is unknown but is not b22

we get:

Card 1 is b but is not 12
Card 2 is 2 but is not 1b
Card 3 is unknown but is not 12b
Card 4 is unknown but is not 2b

Additionally, the input prompt should always display the options in order.

Finally, some style changes in defining a constant for RAINBOW='?',
and in using some more pythonic idioms.