Remove hand-rolled implementation of max and min
Minor changes to hat and cheater to use built-in max and min functions.
This changes the find_highest and find_lowest functions to only return
a single value.  In the case of hat player, only the first value was
ever being used, so the exact behavior is maintained, as established
with /test/regression.py.  For cheater, the random selection is done
differently now, but the output passes manual inspection.