When failing to import players, only fail if they are used
For those who don't have numpy installed, the Encoding players will
fail to import.  When that happens, continue to load the remaining
players, and then let hanabi_wrapper.py handle failure if those players
are requested.  An error message is printed so that the absense of
expected players will not be a complete mystery.
Also change a print statement in heuristics player to be python 3
compatible.  That seems better than catching SyntaxError along with
ImportError.